Fixed tests on Plone 5.1 and lower.
In the export tests, we click to download a json file as attachment.
In Plone 5.2, the json is put in browser.contents.
In Plone 5.1 (Zope 2) and lower, browser.contents is empty.
It works fine when you try it for real though.
I added a patch to catch the data when it is stored on the response,
and save it in a local list.  Ugly, but it gets the job done.