Prevent creating content in a different Plone Site in the same database.
In general, cleanup parent paths when in development on localhost.

Also, in test_import_content_document test what happens when importing twice.
Biggest thing I wanted to check, because I had my doubts:
do we get the original UID on the first import, and a new UID on the second?
Answer: yes.