On Py 2.7, depend on 'jsonschema < 4'.
Version 4 was released in September, and it is Py3-only.
It is pulled in by plone.schema, which we get via plone.restapi.

Note: wedo not use environment markers to specify different dependencies.
This needs setuptools 36.2+, and we still want to support old Plone versions which may use older setuptools.