Fixed ImportError when z3c.relationfield is missing.
I already fixed this half a year or so ago in export_content.py, but now a similar problem is there in export_other.py.
I tried creating an extra tox env for this, with Products.CMFPlone 4.3 instead of Plone,
but z3c.relationfield was pulled in anyway, probably via a test dependency.